Unify filter command file structure
JIRA: RAIL-3466
